+++ kp/cmd/vm/clone.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/luthermonson/go-proxmox"
 	"github.com/spf13/cobra"
-	"github.com/tuana9a/kp/util"
+	"github.com/tuana9a/kp/kp/util"
 )
 
 var cloneCmd = &cobra.Command{
