+++ kp/constants/vm.go
