+++ kp/templates/kube-vip.yaml
