+++ kp/templates/kubevip.go
