+++ kp/templates/worker-setup-1.29.sh
