+++ kp/templates/worker-setup-1.30.sh
