+++ main.go
@@ -1,6 +1,6 @@
 package main
 
-import "github.com/tuana9a/kp/cmd"
+import "github.com/tuana9a/kp/kp/cmd"
 
 func main() {
 	cmd.Execute()
